增加日志打印处理

+++ src/main/java/com/mine/boot/dao/UserMapper.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface UserMapper {
     int deleteByPrimaryKey(Long id);
 
-    int insert(User record);
+    Long insert(User record);
 
     int insertSelective(User record);
